Improve naming of methods in the library
Method naming like "handle_*" imply that the methods are part of an
internal set of methods for responding to events.  In the case of the
CLI form of Esthri they handle subcommands from the CLI, but when
used as a library the naming is awkward because you're not handling
the s3 command, but rather asking esthri to performan an action.