[DEVINFRA-512] Provide a S3 compatibility mode in the CLI
Provides a mode in which the esthri tool can be invoked as though it was
the `aws s3` tool, for select cp/sync operations. Also allows esthri to
serve as a wrapper for the `aws` tool and intercept only commands which
is knows how to handle, transparently handing other commands to the aws
tool.