[DEVINFRA-512] Provide a S3 compatibility mode in the CLI (#207)
* [DEVINFRA-512] Provide a S3 compatibility mode in the CLI

Provides a mode in which the esthri tool can be invoked as though it was
the `aws s3` tool, for select cp/sync operations. Also allows esthri to
serve as a wrapper for the `aws` tool and intercept only commands which
is knows how to handle, transparently handing other commands to the aws
tool.

* Remove aws.esthri python script

Replaced with esthri tool itself now

* Rename SyncParam to S3PathParam

* Separate out the AWS and Esthri CLIs

Put the AWS CLI options into their own struct so that they're separate
from the Esthri options. The benefit of this is that there isn't any
possibility of (unwanted) collisions between Esthri and AWS CLI options.

* Fix build issues

* Add environment var for specifying path to aws

Allows users to specify the path to the aws tool, rather than having it
always hardcoded as aws.real.

* Handle aws s3 options used by Wolf

Handles --quiet (noop), --acl bucket-owner-full-control (noop),
and --include and --exclude for sync.