File browser: Allow display of compressed files [DEVINFRA-513]
Allows the file browser to display gzip compressed files, through
serving them with the content-encoding set to gzip. The way this works
is:

- Files are detected as gzip compressed by searching for the '.gz'
  suffix on the file
- For gzip compressed files, the browser will add links to the 'non
  compressed' (sans .gz) file path with a 'from_gz_compressed' query
  param
- With the 'from_gz_compressed' query param selected, the file browser
  will get the file with gz suffix and serve it.

Also added is the ability to automatically fall back to serving a .gz
file if the path is not available. This is useful for users of esthri
who have uploaded files that have transparently been compressed (and had
a .gz suffix added) so that they can still embed links between files.