aws compat mode: Allow enabling compression
Allows compression to be turned on when esthri is used in the aws
compatibility mode for `s3 sync` commands. This can be turned on
either by setting the `ESTHRI_AWS_COMPAT_MODE_COMPRESSION` environment
variable or by passing the `--compress` command line mode.

This allows (with #226) for 'transparent' compression for users using the
aws compatibility mode. The idea being that users can sync files then
sync them back down again without knowing that they were compressed at
rest in S3. The one drawback of this is that syncing up with compression
currently causes local files to be compressed.