sync: Add a "compressed down" option [DEVINFRA-589] (#226)
* sync: Add a "compressed down" option

Syncing down (from bucket to local path) with compression enabled will
now automatically decompress any (gzip) compressed files. Any files that
are not already compressed will be synced as normal.

Equality checks between remote and local files (which are needed in
order to stop remotely compressed files from being re-downloaded) are
implemented by recompressing local files before syncing. This will incur
a processing cost to save a bandwidth cost. On my laptop, compressing a
1GB random file takes about 30 seconds:

    sam@sambuntu:/tmp$ head -c 1G /dev/urandom > test.random
    sam@sambuntu:/tmp$ time gzip -k test.random

    real        0m27.098s
    user        0m26.474s
    sys         0m0.513s

For my slow internet speeds, this will be better than syncing the whole
file down each time but the equation may be different depending on the
types of files synced, and processing/bandwidth constraints.

If this method of checking for file equality becomes painful for some
use-cases, one way to improve it might be to maintain a cache of md5sums
of uncompressed files to md5sums of compressed files, perhaps in a local
`.esthri` folder. This way, if a particular directory is being regularly
synced then each uncompressed file would not need to be compressed each
time.

* Remove unnecessary path conversion

* Make compressed_path_to_path return Result

Will now return an error if used on a path that is not to a compressed
file.