download: Make directory path if it doesn't exist [DEVINFRA-590]
Allows esthri to work more like `aws s3`, which will create directories
in a download command if they don't exist.

One current difference in this behaviour that would be a future
nice-to-have is that the aws tool only creates directories if the
download is successful. It looks like it downloads to a temp file and
then copies it over.