aws cp compat: Add fallback to compressed file
Adds a fallback to a compressed version of a file (with a '.gz' suffix)
when the AWS compatibility mode is used with compression enabled.

This means that users can use the compatibility mode without realising
it is turned on for eg:

    export ESTHRI_AWS_COMPAT_MODE=1
    export ESTHRI_AWS_COMPAT_MODE_COMPRESSION=1

    aws cp myfile.txt s3://mybucket/ # Will get uploaded as myfile.txt.gz
    aws cp s3://mybucket/myfile.txt . # esthri will fallback to downloading and decompressing to myfile.txt

As part of this, I've also enabled compression mode to work for `aws cp`
commands, in the same way it works for `sync` commands currently.