aws sync: Align filtering to aws tool [SITL-362]
The official `aws` tool's `sync` command implements the `--include` and
`--exclude` filters in a somewhat surprising way, which is different to
how esthri currently handles them in the following two ways:

1. In `aws`, the filters are evaluated from right to left (as they
   appear in the CLI arguments) for each file. The first matching filter
   (either exclude or include) is taken as the decision for that file.
   Currently esthri runs through excludes before includes instead.

2. In `aws`, there is always an implicit `--include *` as the last
   filter. This means `--include` statements are only useful if there is
   a `--exclude` filter. Currently esthri will only select a file if it
   matches an explicit include.

To align the `aws` compatibility mode with the real `aws` tool, this
change makes esthri follow these points by slightly changing the
interface to the `sync` function. API users of this function will now
need to provide a slice of `GlobFilter` enums, where each glob is
specified as either `Include` or `Exclude` and where the order in the
slice determines the order that the filter will be evaluated. Existing
users (just SITL, I think) should be able to retain the same functional
behavior by having this slice be `[exclude filters, include filters]`.

For the second point above, this is implemented at the CLI interface as
to not break existing users of the `sync` API, that might rely on only
explicitly included files being selected.