Make compression feature always on
The compression feature flag adds a bit of verbosity for what is only a
small dependency. This changes removes the feature flag so that flate2
(the compression library) is always included. This should allow the code
to be a little more maintainable at the cost of slightly longer compile
times and slightly larger binary sizes for users compiling with
compression off.