http server: Automatically set content encoding for compressed files
Changes the http server to automatically set the content encoding to
gzip if the file is "esthri compressed" (as determined by looking at the
metadata).

This is a change from the previous way of using a query parameter to
determine whether to serve the compressed file with or without the
content encoding header. The main reason for doing this is that now that
the compression state of a file is stored as metadata rather than as a
file extension, looking up whether a file is compressed ahead of time
requires making a HEAD request. For listing all files this causes a lot
of requests and slows load time.