Transparent compression using metadata [DEVINFRA-594] [DEVINFRA-595] [DEVINFRA-596] [DEVINFRA-597] (#238)
* Make compression feature always on

The compression feature flag adds a bit of verbosity for what is only a
small dependency. This changes removes the feature flag so that flate2
(the compression library) is always included. This should allow the code
to be a little more maintainable at the cost of slightly longer compile
times and slightly larger binary sizes for users compiling with
compression off.

* upload: Don't add .gz suffix in compression mode

Don't modify the filename of uploaded files in compression mode, so that
they can be downloaded later with the same name. We can now rely on the
metadata to tell if a file has been compressed by esthri.

* Move towards enabling transparent compression

* Change CLI args from "compress" to "transparent_compress"

For commands that might involve downloading files (cp & sync), change
the CLI args from "compress" or "decompress" to "transparent_compress",
to signify that these don't necessarily decompress EVERY file, only
files that are actually compressed.

* http server: Automatically set content encoding for compressed files

Changes the http server to automatically set the content encoding to
gzip if the file is "esthri compressed" (as determined by looking at the
metadata).

This is a change from the previous way of using a query parameter to
determine whether to serve the compressed file with or without the
content encoding header. The main reason for doing this is that now that
the compression state of a file is stored as metadata rather than as a
file extension, looking up whether a file is compressed ahead of time
requires making a HEAD request. For listing all files this causes a lot
of requests and slows load time.

* http server: Add .gz suffix to compressed files in archive

When downloading an archive, it can be confusing if files are shown
without any indication that they are compressed. To make this a little
nicer, this adds a ".gz" suffix to the compressed file within the
archive.

* Add a few more tests

* Add AWS fallback "escape hatch"

Allow setting an env var to enable always falling back to the real AWS
tool. This will be useful to set in SITL where we can set the env var in
the container and then override it in some workflows, allowing a slower
rollout.

* Use Cows to avoid string allocation

* Process head requests in parallel

* Head request out of loop

* Fix up whitespace added from removing compression feature

* Remove another allocation using Cow