Reduce upload memory usage [DEVINFRA-609]
To have parity with the `aws` tool we currently upload files in 8MiB
chunks, which allows syncs to work consistently between both tools.

However, currently, this causes a large amount of memory use as reads
are read into 8MiB buffers and with a high amount of upload tasks, this
causes large amounts of data to sit waiting in memory to be uploaded.

This change moves to giving a stream that the rustoto upload function
consumes, that means only 1MiB should be read at a time. This reduces
the amount of memory use, especially when esthri is configured to use a
large amount of upload tasks.