Change retry termination to be based on number of attempts
Changes the retry termination condition to be based on the number of
attempts, rather than the total elapsed time. This works better if the
connection is saturated and requests are taking a longer amount of time.
Without this change, requests were only getting a single attempt before
being failed because they were taking longer than the max total time.