Update default task amounts
Updates the default settings for tasks to work for a lower denominator
(in terms of bandwidth). Having a large amount of connections (up to 512
for a download) on relatively low bandwidth connections (> 50Mbps) was
causing issues with connection resets.

Users of esthri with higher amounts of bandwidth may need to increase
these defaults.