Fix compressed temp files being uploaded [DEVINFRA-617]
Adds a check to make sure compressed temp files don't get uploaded as
part of a sync up. This was causing issues as the temp files were being
created while esthri was running, walkdir was yielding them but then
when it went to access them they'd sometimes already been deleted.

We shouldn't be uploading these files anyway, so I've added a filter to
stop them being yielded in the first place.

Probably weren't seeing this before as this will only show up when the
amount of files being uploaded is much larger than the amount of upload
tasks.