Improvements for low bandwidth connections [DEVINFRA-616] (#255)
* Change retry termination to be based on number of attempts

Changes the retry termination condition to be based on the number of
attempts, rather than the total elapsed time. This works better if the
connection is saturated and requests are taking a longer amount of time.
Without this change, requests were only getting a single attempt before
being failed because they were taking longer than the max total time.

* Update default task amounts

Updates the default settings for tasks to work for a lower denominator
(in terms of bandwidth). Having a large amount of connections (up to 512
for a download) on relatively low bandwidth connections (> 50Mbps) was
causing issues with connection resets.

Users of esthri with higher amounts of bandwidth may need to increase
these defaults.

* Tidy up config

* Fix unknown not being transient

* Update retry log messages

* Remove error log

* Formatting

* Fix compressed temp files being uploaded [DEVINFRA-617]

Adds a check to make sure compressed temp files don't get uploaded as
part of a sync up. This was causing issues as the temp files were being
created while esthri was running, walkdir was yielding them but then
when it went to access them they'd sometimes already been deleted.

We shouldn't be uploading these files anyway, so I've added a filter to
stop them being yielded in the first place.

Probably weren't seeing this before as this will only show up when the
amount of files being uploaded is much larger than the amount of upload
tasks.