chore - update a lot of dependencies (#683)
It's been awhile since we have upgraded dependencies. This PR is mostly
just minor bumps to our dependencies.

However there are a couple breaking changes:
1) The update of reqwest to 0.12 required the headers and headermap to
be from http and not hyper. I just decided to export http in the
esthri-internals crate. I'm not actually sure if that is the right call.

2) When I bumped the MSRV to 1.74 i forgot to change the version in the
`clippy.toml` file. That is unrelated to this change but just saw the
warning when I was running `cargo make lint` locally.